+++ conda/conda-reqs.txt
@@ -1,9 +1,7 @@
-adal>=1.2.2
 attrs>=18.2.0
 azure-common>=1.1.18
 azure-mgmt-network>=2.7.0
 azure-mgmt-resource>=2.2.0
-beautifulsoup4>=4.6.3
 bokeh>=1.4.0
 cryptography>=3.1
 deprecated>=1.2.4
@@ -12,6 +10,7 @@ folium>=0.9.0
 ipython>=7.1.1
 ipywidgets>=7.4.2
 keyring>=13.2.1
+lxml
 matplotlib>=3.0.0
 msrest>=0.6.0
 msrestazure>=0.6.0
@@ -24,7 +23,6 @@ pyyaml>=3.13
 requests>=2.21.1
 scikit-learn>=0.20.2
 scipy
-seaborn>=0.9.0
 setuptools>=40.6.3
 statsmodels
 tqdm>=4.36.1
