+++ docs/source/DataAcquisition.rst
@@ -9,7 +9,7 @@ Querying and Importing Data
    data_acquisition/DataQueries
    data_acquisition/CollectingLinuxAuditLogs
    data_acquisition/UploadData
-   data_acquisition/DataObfuscation
+   data_acquisition/DataMasking
    data_acquisition/MordorData
    data_acquisition/AzureSentinel
    data_acquisition/AzureBlobStorage
