+++ docs/source/msticpy.analysis.rst
@@ -29,3 +29,29 @@ msticpy.analysis.timeseries
     :undoc-members:
     :show-inheritance:
 
+
+msticpy.analysis.eventcluster module
+------------------------------------
+
+.. automodule:: msticpy.sectools.eventcluster
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+msticpy.analysis.outliers module
+--------------------------------
+
+.. automodule:: msticpy.analysis.outliers
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+msticpy.analysis.cluster\_auditd module
+---------------------------------------
+
+.. automodule:: msticpy.analysis.cluster_auditd
+    :members:
+    :undoc-members:
+    :show-inheritance:
