+++ msticpy/data/__init__.py
@@ -1,7 +1,8 @@
 """Data sub-package."""
+from ..common.exceptions import MsticpyImportExtraError
+
 # flake8: noqa: F403
 from .data_providers import QueryProvider
-from .azure_data import AzureData
 from .query_defns import DataEnvironment, DataFamily
 
 from .._version import VERSION
