+++ requirements.txt
@@ -1,18 +1,18 @@
-adal>=1.2.2  # Kql
+# adal>=1.2.2  # Kql
 attrs>=18.2.0
 azure-common>=1.1.18
 azure-core>=1.2.2
 azure-identity==1.4.0
-azure-keyvault-secrets>=4.0.0  # azure
-azure-mgmt-compute>=4.6.2  # azure
-azure-mgmt-core>=1.2.1  # installed by others
-azure-mgmt-keyvault>=2.0.0  # azure
-azure-mgmt-monitor>=1.0.1  # azure
-azure-mgmt-network>=2.7.0  # azure
-azure-mgmt-resource>=2.2.0  # azure
+# azure-keyvault-secrets>=4.0.0  # azure
+# azure-mgmt-compute>=4.6.2  # azure
+# azure-mgmt-core>=1.2.1  # installed by others
+# azure-mgmt-keyvault>=2.0.0  # azure
+# azure-mgmt-monitor>=1.0.1  # azure
+# azure-mgmt-network>=2.7.0  # azure
+# azure-mgmt-resource>=2.2.0  # azure
 azure-mgmt-subscription>=0.2.0
-azure-storage-blob>=12.5.0  #
-beautifulsoup4>=4.6.3  # move to dev
+# azure-storage-blob>=12.5.0  #
+# beautifulsoup4>=4.6.3  # move to dev
 bokeh>=1.4.0
 cryptography>=3.1
 deprecated>=1.2.4
@@ -23,11 +23,12 @@ html5lib  # req by pandas for scraping
 ipwhois>=1.1.0
 ipython>=7.1.1
 ipywidgets>=7.4.2
-keyring>=13.2.1  # azure
-Kqlmagic>=0.1.106
+# keyring>=13.2.1  # azure
+# Kqlmagic>=0.1.106
+lxml
 matplotlib>=3.0.0
 msrest>=0.6.0
-msrestazure>=0.6.0  # azure
+# msrestazure>=0.6.0  # azure
 networkx>=2.2
 numpy>=1.15.4  # pandas
 pandas>=0.25.0
@@ -35,12 +36,12 @@ python-dateutil>=2.8.1  # pandas
 pytz>=2019.2  # pandas
 pyyaml>=3.13
 requests>=2.21.1
-scikit-learn>=0.20.2  # scikit
-scipy>=1.1.0  # timeseries
-seaborn>=0.9.0
+# scikit-learn>=0.20.2  # scikit
+# scipy>=1.1.0  # timeseries
+# seaborn>=0.9.0
 setuptools>=40.6.3
-splunk-sdk>=1.6.0  # splunk
-statsmodels>=0.11.1  # timeseries
+# splunk-sdk>=1.6.0  # splunk
+# statsmodels>=0.11.1  # timeseries
 tldextract>=2.2.2
 tqdm>=4.36.1
 urllib3>=1.23
