+++ tests/data/browsers/test_mordor_browser.py
@@ -17,9 +17,9 @@ _NB_FOLDER = "docs/notebooks"
 _NB_NAME = "MordorData.ipynb"
 
 
-# @pytest.mark.skipif(
-#     not os.environ.get("MSTICPY_TEST_NOSKIP"), reason="Skipped for local tests."
-# )
+@pytest.mark.skipif(
+    not os.environ.get("MSTICPY_TEST_NOSKIP"), reason="Skipped for local tests."
+)
 def test_mordor_browser():
     """Mordor browser UI Test."""
     nb_path = Path(_NB_FOLDER).joinpath(_NB_NAME)
