+++ tests/data/drivers/test_mordor_driver.py
@@ -4,9 +4,7 @@
 # license information.
 # --------------------------------------------------------------------------
 """Mordor data driver test."""
-import contextlib
 from datetime import datetime
-import io
 import os
 from pathlib import Path
 
@@ -28,7 +26,7 @@ _SAVE_PATH = ""
 _SAVE_FOLDER = "mordor_test"
 
 
-@pytest.fixture(scope="module")
+@pytest.fixture(scope="session")
 def qry_provider():
     """Query Provider fixture."""
     Path(_SAVE_FOLDER).mkdir(exist_ok=True)
@@ -105,24 +103,6 @@ def test_mordor_search(mdr_driver: MordorDriver):
 @pytest.mark.skipif(
     not os.environ.get("MSTICPY_TEST_NOSKIP"), reason="Skipped for local tests."
 )
-def test_mordor_download(mdr_driver: MordorDriver):
-    """Test file download."""
-    global _SAVE_PATH
-    entry_id = list(mdr_driver.mordor_data.keys())[2]
-    entry = mdr_driver.mordor_data[entry_id]
-    files = entry.get_file_paths()
-
-    file_path = files[0]["file_path"]
-    d_frame = download_mdr_file(file_path, save_folder="mordor_test")
-    _SAVE_PATH = file_path.split("/")[-1]
-
-    check.is_instance(d_frame, pd.DataFrame)
-    check.greater_equal(len(d_frame), 10)
-
-
-@pytest.mark.skipif(
-    not os.environ.get("MSTICPY_TEST_NOSKIP"), reason="Skipped for local tests."
-)
 def test_mordor_query_provider(qry_provider):
     """Test query functions from query provider."""
     queries = qry_provider.list_queries()
@@ -132,13 +112,10 @@ def test_mordor_query_provider(qry_provider):
     check.is_true(hasattr(qry_provider, queries[0]))
 
     q_func = getattr(qry_provider, queries[2])
-    output = io.StringIO()
-    with contextlib.redirect_stdout(output):
-        q_func("?")
-    check.is_in("Query:", output.getvalue())
-    check.is_in("Data source:  Mordor", output.getvalue())
-    check.is_in("Mordor ID:", output.getvalue())
-    check.is_in("Mitre Techniques:", output.getvalue())
+
+    check.is_in("Notes", q_func.__doc__)
+    check.is_in("Mordor ID:", q_func.__doc__)
+    check.is_in("Mitre Techniques:", q_func.__doc__)
 
     f_path = q_func("print")
     check.is_in("https://raw.githubusercontent.com/OTRF/mordor", f_path)
