+++ tests/test_auditdextract.py
@@ -9,16 +9,26 @@ import unittest
 import os
 
 import pandas as pd
+import pytest
+import pytest_check as check
 
 from msticpy.sectools.auditdextract import (
     extract_events_to_df,
     get_event_subset,
     generate_process_tree,
-    cluster_auditd_processes,
     read_from_file,
 )
 
 from msticpy.sectools.process_tree_utils import get_summary_info
+
+_CLUSTER_OK = False
+try:
+    from msticpy.analysis.cluster_auditd import cluster_auditd_processes
+
+    _CLUSTER_OK = True
+except ImportError:
+    pass
+
 
 _test_data_folders = [
     d for d, _, _ in os.walk(os.getcwd()) if d.endswith("/tests/testdata")
@@ -119,12 +129,18 @@ class TestAuditdExtract(unittest.TestCase):
         self.assertGreaterEqual(len(proc_tree), 85)
         self.assertEqual(pt_summary["LargestTreeDepth"], 5)
 
-        proc_events2 = proc_events.copy()
-        clustered_procs = cluster_auditd_processes(proc_events2, app=None)
-        self.assertIsNotNone(clustered_procs)
-        self.assertEqual(len(clustered_procs), 2)
 
+@pytest.mark.skipif(not _CLUSTER_OK, reason="Partial install of msticpy")
+def test_auditd_cluster():
+    input_file = os.path.join(_TEST_DATA, "linux_events.csv")
+    input_df = pd.read_csv(input_file)
 
-if __name__ == "__main__":
-    unittest.main()
-    print("bye")
+    input_df["AuditdMessage"] = input_df.apply(
+        lambda x: ast.literal_eval(x.AuditdMessage), axis=1
+    )
+    output_df = extract_events_to_df(data=input_df)
+    proc_events = get_event_subset(output_df, event_type="SYSCALL_EXECVE")
+
+    clustered_procs = cluster_auditd_processes(proc_events, app=None)
+    check.is_not_none(clustered_procs)
+    check.equal(len(clustered_procs), 2)
