+++ tests/test_security_event.py
@@ -5,17 +5,10 @@
 # --------------------------------------------------------------------------
 """Security Event test class."""
 import unittest
-import json
 import os
 
 import pandas as pd
 
-from msticpy.sectools.eventcluster import *
-from msticpy.sectools.eventcluster import (
-    token_count_df,
-    delim_count_df,
-    char_ord_score_df,
-)
 from msticpy.nbtools.security_event import SecurityEvent
